requirements.txt
# 聊天记录生成器依赖包

# Google AI API (用于AI聊天生成功能)
google-generativeai>=0.3.0

# 环境变量管理
python-dotenv>=0.19.0

# 其他依赖 (如果需要的话)
# requests>=2.25.0
